@@ -1,6 +1,6 @@
 {
   "name": "parse-server",
-  "version": "2.2.19",
+  "version": "2.2.23",
   "description": "An express module providing a Parse-compatible API server",
   "main": "lib/index.js",
   "repository": {
@@ -18,62 +18,61 @@
   ],
   "license": "BSD-3-Clause",
   "dependencies": {
-    "babel-polyfill": "6.13.0",
     "bcryptjs": "2.3.0",
     "body-parser": "1.15.2",
     "commander": "2.9.0",
     "deepcopy": "0.6.3",
     "express": "4.14.0",
     "intersect": "1.0.1",
-    "lodash": "4.15.0",
+    "lodash": "4.16.4",
     "lru-cache": "4.0.1",
     "mime": "1.3.4",
-    "mongodb": "2.2.9",
+    "mongodb": "2.2.10",
     "multer": "1.2.0",
-    "parse": "1.9.1",
+    "parse": "1.9.2",
     "parse-server-fs-adapter": "1.0.1",
     "parse-server-push-adapter": "1.1.0",
     "parse-server-s3-adapter": "1.0.5",
     "parse-server-simple-mailgun-adapter": "1.0.0",
-    "pg-promise": "5.3.3",
+    "pg-promise": "5.3.5",
     "redis": "2.6.2",
-    "request": "2.74.0",
+    "request": "2.76.0",
     "semver": "5.2.0",
     "tv4": "1.2.7",
     "winston": "2.2.0",
-    "winston-daily-rotate-file": "1.3.0",
+    "winston-daily-rotate-file": "1.3.1",
     "ws": "1.1.1"
   },
   "devDependencies": {
-    "babel-cli": "6.14.0",
-    "babel-core": "6.14.0",
+    "babel-cli": "6.18.0",
+    "babel-core": "6.18.0",
     "babel-plugin-syntax-flow": "6.13.0",
-    "babel-plugin-transform-flow-strip-types": "6.14.0",
+    "babel-plugin-transform-flow-strip-types": "6.18.0",
     "babel-preset-es2015": "6.14.0",
     "babel-preset-stage-0": "6.5.0",
-    "babel-register": "6.14.0",
+    "babel-register": "6.18.0",
     "bcrypt-nodejs": "0.0.3",
-    "cross-env": "2.0.1",
+    "cross-env": "3.1.3",
     "deep-diff": "0.3.4",
     "gaze": "1.1.1",
     "istanbul": "1.0.0-alpha.1",
-    "jasmine": "2.4.1",
+    "jasmine": "2.5.2",
     "mongodb-runner": "3.4.0",
-    "nodemon": "1.10.0",
+    "nodemon": "1.11.0",
     "request-promise": "4.1.1"
   },
   "scripts": {
     "dev": "npm run build && node bin/dev",
     "build": "babel src/ -d lib/",
-    "test": "cross-env MONGODB_VERSION=${MONGODB_VERSION:=3.2.6} MONGODB_STORAGE_ENGINE=mmapv1 NODE_ENV=test TESTING=1 babel-node $COVERAGE_OPTION ./node_modules/jasmine/bin/jasmine.js",
-    "test:win": "npm run pretest && cross-env NODE_ENV=test TESTING=1 babel-node ./node_modules/jasmine/bin/jasmine.js && npm run posttest",
+    "test": "cross-env MONGODB_VERSION=${MONGODB_VERSION:=3.2.6} MONGODB_STORAGE_ENGINE=mmapv1 NODE_ENV=test TESTING=1 node $COVERAGE_OPTION ./node_modules/jasmine/bin/jasmine.js",
+    "test:win": "npm run pretest && cross-env NODE_ENV=test TESTING=1 node ./node_modules/jasmine/bin/jasmine.js && npm run posttest",
     "coverage": "cross-env COVERAGE_OPTION='./node_modules/.bin/istanbul cover' npm test",
-    "coverage:win": "npm run pretest && cross-env NODE_ENV=test TESTING=1 babel-node ./node_modules/babel-istanbul/lib/cli.js cover ./node_modules/jasmine/bin/jasmine.js && npm run posttest",
+    "coverage:win": "npm run pretest && cross-env NODE_ENV=test TESTING=1 node ./node_modules/babel-istanbul/lib/cli.js cover ./node_modules/jasmine/bin/jasmine.js && npm run posttest",
     "start": "node ./bin/parse-server",
     "prepublish": "npm run build"
   },
   "engines": {
-    "node": ">=4.3"
+    "node": ">=4.5"
   },
   "bin": {
     "parse-server": "./bin/parse-server"
