@@ -95,5 +95,5 @@ export function removeTransport(transport) {
   });
 }
 
-export { logger, addTransport, configureLogger, removeTransport };
+export { logger };
 export default logger;
