@@ -1,7 +1,7 @@
 
 import PromiseRouter from '../PromiseRouter';
 import rest          from '../rest';
-
+import _             from 'lodash';
 import url           from 'url';
 
 const ALLOWED_GET_QUERY_KEYS = ['keys', 'include'];
@@ -115,7 +115,7 @@ export class ClassesRouter extends PromiseRouter {
 
   static JSONFromQuery(query) {
     let json = {};
-    for (let [key, value] of Object.entries(query)) {
+    for (let [key, value] of _.entries(query)) {
       try {
         json[key] = JSON.parse(value);
       } catch (e) {
