@@ -1,43 +1,13 @@
-function numberParser(key) {
-  return function(opt) {
-    opt = parseInt(opt);
-    if (!Number.isInteger(opt)) {
-      throw new Error(`The ${key} is invalid`);
-    }
-    return opt;
-  }
-}
+import {
+  numberParser,
+  numberOrBoolParser,
+  objectParser,
+  arrayParser,
+  moduleOrObjectParser,
+  booleanParser,
+  nullParser
+} from '../utils/parsers';
 
-function objectParser(opt) {
-  if (typeof opt == 'object') {
-    return opt;
-  }
-  return JSON.parse(opt)
-}
-
-function moduleOrObjectParser(opt) {
-  if (typeof opt == 'object')  {
-    return opt;
-  }
-  try {
-    return JSON.parse(opt);
-  } catch(e) {}
-  return opt;
-}
-
-function booleanParser(opt) {
-  if (opt == true || opt == "true" || opt == "1") {
-    return true;
-  }
-  return false;
-}
-
-function nullParser(opt) {
-  if (opt == 'null') {
-    return null;
-  }
-  return opt;
-}
 
 export default {
   "appId": {
@@ -119,9 +89,7 @@ export default {
     env: "PARSE_SERVER_FACEBOOK_APP_IDS",
     help: "Comma separated list for your facebook app Ids",
     type: "list",
-    action: function(opt) {
-      return opt.split(",")
-    }
+    action: arrayParser
   },
   "enableAnonymousUsers": {
     env: "PARSE_SERVER_ENABLE_ANON_USERS",
@@ -229,6 +197,29 @@ export default {
   },
   "cluster": {
     help: "Run with cluster, optionally set the number of processes default to os.cpus().length",
-    action: numberParser("cluster"),
-  }
+    action: numberOrBoolParser("cluster")
+  },
+   "liveQuery": {
+    help: "parse-server's LiveQuery configuration object",
+    action: objectParser
+  },
+  "liveQuery.classNames": {
+    help: "parse-server's LiveQuery classNames",
+    action: arrayParser
+  },
+  "liveQuery.redisURL": {
+    help: "parse-server's LiveQuery redisURL",
+  },
+  "startLiveQueryServer": {
+    help: "Starts the liveQuery server",
+    action: booleanParser
+  },
+  "liveQueryPort": {
+    help: 'Specific port to start the live query server',
+    action: numberParser("liveQueryPort")
+  },
+  "liveQueryServerOptions": {
+    help: "Live query server configuration options (will start the liveQuery server)",
+    action: objectParser
+  },
 };
