@@ -1,18 +1,12 @@
 import path from 'path';
 import express from 'express';
 import { ParseServer } from '../index';
-import definitions from './cli-definitions';
-import program from './utils/commander';
-import { mergeWithOptions } from './utils/commander';
+import definitions from './definitions/parse-server';
 import cluster from 'cluster';
 import os from 'os';
+import runner from './utils/runner';
 
-program.loadDefinitions(definitions);
-
-program
-  .usage('[options] <path/to/configuration.json>');
-
-program.on('--help', function(){
+const help = function(){
   console.log('  Get Started guide:');
   console.log('');
   console.log('    Please have a look at the get started guide!')
@@ -32,33 +26,7 @@ program.on('--help', function(){
   console.log('    $ parse-server -- --appId APP_ID --masterKey MASTER_KEY --serverURL serverURL');
   console.log('    $ parse-server -- --appId APP_ID --masterKey MASTER_KEY --serverURL serverURL');
   console.log('');
-});
-
-program.parse(process.argv, process.env);
-
-let options = program.getOptions();
-
-if (!options.serverURL) {
-  options.serverURL = `http://localhost:${options.port}${options.mountPath}`;
-}
-
-if (!options.appId || !options.masterKey || !options.serverURL) {
-  program.outputHelp();
-  console.error("");
-  console.error('\u001b[31mERROR: appId and masterKey are required\u001b[0m');
-  console.error("");
-  process.exit(1);
-}
-
-function logStartupOptions(options) {
-  for (let key in options) {
-    let value = options[key];
-    if (key == "masterKey") {
-      value = "***REDACTED***";
-    }
-    console.log(`${key}: ${value}`);
-  }
-}
+};
 
 function startServer(options, callback) {
   const app = express();
@@ -66,7 +34,15 @@ function startServer(options, callback) {
   app.use(options.mountPath, api);
 
   var server = app.listen(options.port, callback);
-
+  if (options.startLiveQueryServer || options.liveQueryServerOptions) {
+    let liveQueryServer = server;
+    if (options.liveQueryPort) {
+      liveQueryServer = express().listen(options.liveQueryPort, () => {
+        console.log('ParseLiveQuery listening on ' + options.liveQueryPort);
+      });
+    }
+    ParseServer.createLiveQueryServer(liveQueryServer, options.liveQueryServerOptions);
+  }
   var handleShutdown = function() {
     console.log('Termination signal received. Shutting down.');
     server.close(function () {
@@ -77,27 +53,59 @@ function startServer(options, callback) {
   process.on('SIGINT', handleShutdown);
 }
 
-if (options.cluster) {
-  const numCPUs = typeof options.cluster === 'number' ? options.cluster : os.cpus().length;
-  if (cluster.isMaster) {
-    logStartupOptions(options);
-    for(var i = 0; i < numCPUs; i++) {
-      cluster.fork();
+
+runner({
+  definitions,
+  help,
+  usage: '[options] <path/to/configuration.json>',
+  start: function(program, options, logOptions) {
+    if (!options.serverURL) {
+      options.serverURL = `http://localhost:${options.port}${options.mountPath}`;
     }
-    cluster.on('exit', (worker, code, signal) => {
-      console.log(`worker ${worker.process.pid} died... Restarting`);
-      cluster.fork();
-    });
-  } else {
-    startServer(options, () => {
-      console.log('['+process.pid+'] parse-server running on '+options.serverURL);
-    });
+
+    if (!options.appId || !options.masterKey || !options.serverURL) {
+      program.outputHelp();
+      console.error("");
+      console.error('\u001b[31mERROR: appId and masterKey are required\u001b[0m');
+      console.error("");
+      process.exit(1);
+    }
+
+    if (options["liveQuery.classNames"]) {
+      options.liveQuery = options.liveQuery || {};
+      options.liveQuery.classNames = options["liveQuery.classNames"];
+      delete options["liveQuery.classNames"];
+    }
+    if (options["liveQuery.redisURL"]) {
+      options.liveQuery = options.liveQuery || {};
+      options.liveQuery.redisURL = options["liveQuery.redisURL"];
+      delete options["liveQuery.redisURL"];
+    }
+
+    if (options.cluster) {
+      const numCPUs = typeof options.cluster === 'number' ? options.cluster : os.cpus().length;
+      if (cluster.isMaster) {
+        for(var i = 0; i < numCPUs; i++) {
+          cluster.fork();
+        }
+        cluster.on('exit', (worker, code, signal) => {
+          console.log(`worker ${worker.process.pid} died... Restarting`);
+          cluster.fork();
+        });
+      } else {
+        startServer(options, () => {
+          console.log('['+process.pid+'] parse-server running on '+options.serverURL);
+        });
+      }
+    } else {
+      startServer(options, () => {
+        logOptions();
+        console.log('');
+        console.log('['+process.pid+'] parse-server running on '+options.serverURL);
+      });
+    }
   }
-} else {
-  startServer(options, () => {
-    logStartupOptions(options);
-    console.log('');
-    console.log('['+process.pid+'] parse-server running on '+options.serverURL);
-  });
-}
+})
+
+
 
