@@ -1,5 +1,18 @@
 'use strict';
-let logger;
+import defaults from './defaults';
+import { WinstonLoggerAdapter } from './Adapters/Logger/WinstonLoggerAdapter';
+import { LoggerController }     from './Controllers/LoggerController';
+
+function defaultLogger() {
+  let adapter = new WinstonLoggerAdapter({
+    logsFolder: defaults.logsFolder,
+    jsonLogs: defaults.jsonLogs,
+    verbose: defaults.verbose,
+    silent: defaults.silent });
+  return new LoggerController(adapter);
+}
+
+let logger = defaultLogger();
 
 export function setLogger(aLogger) {
   logger = aLogger;
